requirements.txt
absl-py==0.5.0
astor==0.7.1
backports.weakref==1.0.post1
Django==1.11.16
django-cors-headers==2.4.0
django-extra-fields==1.0.0
django-rest-framework==0.1.0
djangorestframework==3.8.2
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
mock==2.0.0
numpy==1.15.2
opencv-python==3.4.3.18
pbr==4.3.0
Pillow==5.3.0
protobuf==3.6.1
pytz==2018.5
six==1.11.0
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
Werkzeug==0.14.1
